cleanup: move g72x stuff around
