+++ src/libvgmstream.vcxproj
@@ -106,7 +106,6 @@
     <ClInclude Include="base\seek_table.h" />
     <ClInclude Include="base\tags.h" />
     <ClInclude Include="coding\coding.h" />
-    <ClInclude Include="coding\g72x_state.h" />
     <ClInclude Include="coding\mpeg_decoder.h" />
     <ClInclude Include="coding\vorbis_custom_decoder.h" />
     <ClInclude Include="coding\libs\binka_data.h" />
@@ -120,6 +119,7 @@
     <ClInclude Include="coding\libs\g7221_aes.h" />
     <ClInclude Include="coding\libs\g7221_data.h" />
     <ClInclude Include="coding\libs\g7221_lib.h" />
+    <ClInclude Include="coding\libs\g72x_vgmstream.h" />
     <ClInclude Include="coding\libs\icelib.h" />
     <ClInclude Include="coding\libs\ka1a_dec.h" />
     <ClInclude Include="coding\libs\ka1a_dec_data.h" />
@@ -372,6 +372,7 @@
     <ClCompile Include="coding\libs\compresswave_lib.c" />
     <ClCompile Include="coding\libs\g7221_aes.c" />
     <ClCompile Include="coding\libs\g7221_lib.c" />
+    <ClCompile Include="coding\libs\g72x_vgmstream.c" />
     <ClCompile Include="coding\libs\icelib.c" />
     <ClCompile Include="coding\libs\ka1a_dec.c" />
     <ClCompile Include="coding\libs\libacm_decode.c" />
